Clean up the help_overview
the .md output only displayed on some markdown viewers.

It was just as easy to change to using .rst as the output so it now
produces .rst output.

Note that this is still completely manual so that to use it you must
enter

./help_overview.py >../design/help_overview.rst.

Goal is to:

1. Integrate the display into the docs.

2. Automate it so that it automatically finds every group and subcommand
similar to how the click man file generator contribution works.