Add a  new help overview program based on the output strucuture of click --help output. This allows us to gather all help information without forgetting if something has changed.  It finds all subcommands and generates a recursive list of all help then displays it in .rst format
Remove bad dir from initial commit. That was an error
in copying the code from another project

Update resulting .rst files in doc and design

Remove original tool

Implement exists test for pywbemcli so that if
the executable does not exist, the program
will exit with an error.  It is then the
problem for the developer to be sure that
pywbemcli exists.