Create a more general table output mechanism
Create a function specifically to print tables in a
variety of formats (plain, simple, grid, etc.)  These
are all ascii table formatters

This function is named format_table and there is a
corresponding folding function (fold_string) that can
fold a long string to create a multi-line cell in a
table.

The format table function uses the terminaltable package
to format the table because today this package supports
multiline cells in tables. A the same time, it uses the
format names defined by tabulate so that at some point
in the future we can move to use tabulate package with
minimal changes.

Added output_format cmd line parameter to pywbemcli and
a list of specific output formats to cover both cim
object and table outputs.

We moved several of the click commands from just using
a raw output lister to using the table formatter to
generate output. Specifically this included the
commands in _cmd_server so that they all output in
a format based on a table with headers and optionally
a title.

Adds unit test for table_format (test_tableformat.py)