# This is a combination of 2 commits.
# The first commit's message is:

Implement operation statistics with display

This implements the capture and display of pywbem operation statistics
for each pywbemcli opertion executed.  It adds a new pywbemcli option
(-T, --time_stats) that turns the statistics on.

This is based on the time statistics added to pywbem in version 0.11.0
(not yet released) and so, for the moment pips pywbem directly from
github and NOT from PIP.

The display appears as a table showing operation count, exception count,
execution time, request and reply lengths, and operation name for all
operations executed for a particular pywbemcli subcommand.

If the wbem server is providing server time, that is also included.

This pr also corrects a couple of errors found in the server group and
changes the form of the display for namespaces and interop to tables.
Note that the current table form is temporary.   Issue #50 proposes a
cleanup of tables.

# This is the 2nd commit message:

Updates documentation and the help details file