Clean up documentation
Clean up the text in the code that is used for help documentation to
limit the summary line to what can be displayed (avoid the ... summary
lines for subcommands) and clarify the functionality.

Also changes name of the hierarchy subcommand to tree as being more
expressive and easier to spell.

Change references one char option names so that references and associators match.