This pr rebased to pr ks/#xxx-create_funct_tests which adds the tests/function directory the utils.py file that provides functions to execute pywbemcli both as a script and as in line code  and some initital tests. The mock tests will be built on that basis to correct variable throughout pywbemcli code.
Addsupport for pywbem_mock

This pr adds support for pywbem_mock by extending the code where a
WBEMConnection is created to create either a mock or real wbem
connection.

The choice of a real or fake connection is determined by the url
scheme on input. The definition of data to be inserted into the
fake repository is determined by the netloc and path components of the
url if the url scheme is 'fake'.

To do this we modified the code that adds specific pywbemcli
methods to wbemconnection to add them as a mixin rather than a subclass
and added code to build the mock repository from the cmd line received.
The url scheme for faking is 'fake'. The netloc is a type of mock.
Today it is only 'file'. The url path for the mock url is the pathname
of the mof file relative to the current directory.

This also fixes an issue where we were using uri for the variable names
and it is really url.  Renamed the variables.

Add TestEnumerate class to test_class_subcmd.py
Also added a first use of the mock mechanism  in TestEnumerate class
with the method test_enumerate_mock. This is a simplistic first
cut and only executes a single test.