Add mock server support to pywbemcli.
1. Created a new global cmd line option that allows defining multiple
filenames that will force the use of the mock support in place of wbem
connection and allow adding mof to the repository or executing
arbitrary python code as part of the startup.  The option is
--mock_server and it can be defined multiple times on a single pywbemcli
call.  Each use of the option defines a single file name or file path
name that can be used to populate the mock repository.

This required changes to the context object and pywbemcli_server to
pass this new method and to make the choce of which connection to which
we connect in pywbemcli_server based on the existence of the mock_server
global option.Because it was proving difficult to subclass WBEMConnection
and FakeWBEMConnection to include the special pywbemcli methods we added
for pywbemcli usage, we changed this code to a mixin rather than a subclass.
We also added a mixin just for the  FakeWBEMConnection that builds the
repository.

2. Created a file in tests/unit (utils.py) that executes pywbemcli script
and returns results with some asserts to test the returns. This includes
asserts associated with the results of executing pywbemcli. This will be
the general pywbemcli script executor in the future.

3. Created a new test file unit/test_global_opts.py in tests/unit to test
the utils.py. Right now it simply tests the basic help and the new
--mock_server option.

4. Created a first subcommand test that tests the basics of the class enumerate
using the mock support.  This confirms that mof can be installed and python
files executed  and also serves as the basis for futher tests.
NOTE: We want to move this to use the new pywbem simple_test_function
when that is committed so the parameters are very basic for the moment.

5. Changed code to use _url in place of _uri in a number of variable
names to be consistent with pywbem. This had been integrated in early
and it was easier to leave it in than back it out. Sorry. Should have left
this one out but it was easier to leave it in than remove it.

6. Expanded the repr of the pywbemcli_server  class for more
complete presentation of the data in the class to help testing the mock
code.

7. Added section to docs documenting the mock support.