Change output table formatting package to use tabulate.
This is a simple change (only a few lines of code) but the tests had to
change because there are subtile changes to the output formatting.  We
also added 'rst' as a possible format and added a test for that format.

We had been using the package terminaltables before because it supported
folded cells but it was a weaker package in general and had always
intended to use tabulate when it incorporated the folded cell logic.
Tests for folded cells have been in the table output format test for
some time now.

While we want to change the tests to pytest, we are limiting this pr to just
the minimum, change package and fix tests that broke because of
formatting differences between terminaltable and tabulate