Increased M2Crypto minimum version to 0.30.1
M2Crypto 0.29.0 introduced support for Windows again, so it is no longer
necessary to use the very outdated M2CryptoWin32/64 packages.

This change uses the M2Crypto package now also on Windows, and eliminates
using M2CryptoWin32/64.

To enable that, it increases the minimum version of M2Crypto to 0.30.1
(which has a minor fix for the Windows support introduced with 0.29.0).

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>