Increased the set of Python envs for Travis
This change increases the number of Python environments on Travis:
- Added Py27 minimum
- Changed Py34 from latest to minimum
- Added Py36 minimum

Reason: It was found that the environment with minimum package level
fail every now and then, and this should not go by undetected.
The manual-ci-run also would catch those, but it is more direct
to catch them in the PRs right away.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>