Refactor the pywbemcli executable tests to create a common base class for the tests  using the mock environment and modify the test for class to use this environment.
All tests are defined through parameters to the single method
that executes the test.

This turned out to be cleaner and more consistent that using
the wrapper function at this point.

The limits with this approach are:

1. The mock environment has to be created for each test which takes
time.

2. Only a single subcommand can be executed within a single test because
the repl today only support entering a single command at at time. Thus
tests like class delete can only confirm a good response, not that the
class was actually deleted in the mock environment.

Note that these tests are not complete today but include a number of
TODOs and test with the condition marked false.

In addition we found a couple of errors in the current code that we fixed
including:

1. On the class tree command, we were not stopping the spinner before
calling the display method.

2. We were using the click form
click.option('--includequalifiers/--no_includequalifiers'
to define the option for include qualifiers.  First it was
giving some problems and second it was just more complex
that we need since the real option is just to not include
qualifiers so I modified that to a simple flag with the name
--no-qualifiers for the getclass, etc. subcommands in the
class group.

3. We had the EnumerateInstances options DeepInheritance
mispelled as Deepinheritance so fixed that both for instance
and class subcommands.

Adds a new file cli_test_extensions.py