Redo mock_server option
Rebased with ks/#117-fix-broken-history to fix history

Redid mock_server option so name is mock-server and there is a short
form.  Added env var feature for this option. Added it to the connection
features so that a new mock with a name can be created. This meant a lot
of renaming in doc to change the _ to - and modification of a couple of
tests.