Update requirements.txt to only use version 1 of prompt-toolkit
Added to help for repl that up and down arrows show history.

This picked up some things that are different in the docs/prompt-
details.

remove other things that are not part of this commit.  Note that we
sometimes pick up individual changes to the help because of the
automatic nature of generating that file.