Fixed issue where help changed on a command
Changed test to match changed help on a command.  Also found that I had
duplicated one help text variable in test_class_subcmd.py

Fixed comments on the generality of some of the options text to clarify
exactly what is being returned. Also tried to clarify what the default
is for booleans with the phrase, if set.

Updates pywbemclicmdshelp.rst also