Add tests for the qualifier group
These tests use calls to pywbemcli to test the help responses of the
group and subcommands as well as the standard mock to test data
responses for each subcommand and its options.

This extends the mock_subcmd_test method to allow for the inclusion of
global command options like --output-format.

It adds a new test file to test the qualifier command group.

It makes one small change to the qualifier subcommands to standardize
the first line of the help text.