Add mock tests for instance subcommand (#110)
* Add mock tests for instance subcommand

Creates a new test and adds tests for the subcommands in the instance
command group.

Adds a single new file with tests for the instance group.  This includes
both valid and error instance tests and includes tests for at
least the following subcommands:

1. enumerate
2. get
3. delete
4. create
5. count
6. references
7. associations

NOTE: Some tests are marked TODO and not done yet.

Add new test scripts and mof for tests

This adds a new mof file for association tests and a new script that
defines a method to be invoked. This new mof file will be retroed
back to the test_class_subcmd.py tests also but in a separate pr

This change also fixes code issues in _cmd_instances.py and
_common.py that were found when we created the tests:

    1. click.echo with missing % in couple of places
    2. variable named classname changed to classname_
    3. the function in common pick_from_list used different
       messages for initial and loop prompts.

* Fixed comments on initial pr.

1. Clarifies help text in several cases, in particular what is returned
in some of the options.

2. Clarified other wording (ex. delete to process because this option
used for multiple subcommands, not just delete)