Adds new cli subcommand to modify instance.
Adds new pywbemcli subcommand modify to _cmd_instance and tests to verify
the new method.

1. Add subcommand for modify_instance.  "pywbemcli instance modify ...".
The command includes the same options as "pywbem instance create" except
it uses instance name as an argument and includes a propertylist option.
This method modifies instances by building a new instance with just
the properties defined in the call.

2. Fix issue in common with resolve_propertylist.  Was testing wrong
part of variable and confusing the case of propertylist = "" (string
empty that should return no proprties)

3. Extended tests for pywbemcli instance create and defined a number of
tests for pywbemcli instance modify

3. Add a verify_operation to instance create and include in instance
modify so that the user can verify if the instance to be created or the
modification is correct before committing to the operation. Displays the
new/modified instance if this option is set and waits for prompt
response from user

4. Modify parse_wbemuri_str to add test for instance name of the form
classname.keybindings without the ":" before the classname.  If this is
found, it adds the classname.  This is a temporary fix until the  issue
in pywbem is fixed and released. The issue behind this is pywbem issue

5. Code cleanup in _common.parse_wbemuri_str function.

6. Change several exceptions in _common to a) drop the word Error at the
beginning of click.ClickException instances (click does that), put ":"
after the word Exception, and clarify wording.

7. Remove a bad test in common.display_cim_objects that was causing a
failure with a single object.

8. Updated the help command info in docs. That is automatic when the
help changes.

9. Cleaned up the text for pywbem instance create