Convert use of tocimobj to cimvalue
Change the use of tocimobj to use the new cimvalue function instead.
Since the new function does not handle the 'True' and 'False' case
we externalized that.

Modified create_cimproperty slightly to avoid passing cim_class to it.

This pr also cvts a number of the tests in test_common.py to pytest
and adds tests for create_ciminstance