Add basic logging to pywbemcli as cmd line option. (#125)
* Add basic logging to pywbemcli as cmd line option.

Adds the --log option. Which accepts a string as input
and passes that string onto pywbem configure_loggers_from_string
function. This means that logging can be defined on the
command line or with an environment variable covering the
various configuration options in pyebem.

There are some changes to _cmd_connection to fix an error in
with misnamed parameter.

Includes test file for log_options that tests configuring and
displaying log.  Note that it only tests log to stderr.

* This also adds a section to the documentation defining logging