Removes remaining use of tocimobj in favor of cimvalue function  and fixes issues in invokemethod
This code rebased with pr # 133 to get other test changes, the
change to pywbem 0.12.6, and the change to pytest-cov versions.

1. Removes the last use of tocimobj which is in the class group for
invokemethod

2. There were no tests for invokemethod so we added them to test the
   new code in test_class_subcmd.  In this case we found some issues
   including:

a. The cli_test extension did not allow multiple mock files to be
   loaded. Extended it to allow multiple files.

b. The mock model had only IN for the simplest method (Fuzzy) and we
   wanted to test both input and output so added OUT as a qualifier on that
   that method in the CIM_Foo class. Also names for methods were poor so
   we made them more meaningful/

c. Added to the description of the methods what we expect it to return.

d. Fixes issue with caps in parameter in definition of create_cimvalue

e. classname and methodname were reversed on call to class level
   InvokeMethod

3. Added invokemethod tests in test_instance_subcommand.py.

4. Added an example to the help documentation for pywbemcli instance
invokemethod.

5. Commonized the code for instance and class invokemethod action
functions (see _common.py:process_invokemethod) since they are the same
except for the input object (class vs. instancename).

6. Modified wording slightly for invokemethod help. This changes
pywbem clicmdshelp.rst