Fix version of pytest-cov to match pywbem and eliminate failure in tests.
Also fixes a couple of places pywbem name used and should be pywbemtools
in requirements.txt, etc.

The following changes gradually grew as we tried to get the code working
again because at the same time as we were working this pr, a new pybem
revision changed some things and broke a number of our tests.

Removed test prints. These were part of temp tests for th ":" header on
from_wbem_uri and got left in by accident. We removed them completely
because we have now concluded that the issue does exist in pywbem 0.12
so we will only remove the code documented in issue #131 when we set
the requirements and minimum requirements to 0.13

Used opportunity to fix test_class_subcmd.py tests. We were failing
because pywbem move to 0.12.6 changed some of error messages
slightly and we were depending on exact messages.

Modified requirements.txt and minimum_constraints.txt to use pywbem
0.12.6 because in the changes from 0.12.5 to 0.12.6 we modified some
messages used in tests which broke a number of our tests.  Fixed both
requirements.txt and minimum_constraints to use 0.12.6.

I used this opportunity to retest several tests in test_class_subcmd.py
that were marked False, fix issues and enable the tests.  We now only
have one test in that group that is marked Fail.

Added the variables OK, RUN, and FAIL to test_class_subcommand.py to make
testing easier.  It already exists in the tests for the other subcommands.

Added one test in test_common.py to cover the issue of the
problem with from_wbemuri that exists with pywbem 0.12, issue #131.

Added one sentence of documentation to cli_test_extenstions.py in
tests/unit because as I changed some of the tests to use regex I found the wording incomplete.