Fix issue in test_instances_subcmd.py causing failue of tests.
There were several cases for the tests are now failing,
notabably a set of changes to capatilize INSTANCENAME in the usage
and add a line to the usage about looking at the help.
This was a problem caused by the fact that we were testing exact
messages and the message from click for missing argument changed between
versions 6.6 and 6.7.

Solution is to reduce the test to just compare the Usage:... part and
ignore the rest of the message.