Fix issue in test_instances_subcmd.py
There were several cases for, whatever the reason, the tests had not
kept  up with the changes in the code, notabably a set of changes to
capatilize INSTANCENAME in the usage and add a line to the usage about
"Try pywbem .... -h".  This simple corrects the text in the test so that
the test passes.