Extend pywbemcli tests to test env variables and stdin input.
In  previous test definitions we ignored the cmd line env variables.
This pr extends the test case definitions to  allow these variables
to be defined as part of tests case definitions.

Since click and click.repl allow the use of stdin to be passed to the
repl and specifically multiline stdin be passed a line at a time we have
a way to test multiple subcommands in a single test.  This pr extends the
test environment to allow single and multiline inputs to be defined as
part of the test setup and with the calls to pywbemcli.

Originally cli_test_extensions.py test class was to be MOCK only but
it has been  generalized to where those tests requiring MOCK define
the mock environment but many of the tests do not use the mock
envirionment. Therefore we removed the MOCK_ from the TEST_CASES name
and the test definitions.

This adds some sample tests to validate the stdin test option and also
refactors some of the test code in cli_test_extensions.py.

Fixed issue in create instance where exception catch incomplete.
It was catching only the CIMError with invalid class and passing
if any other error occurred. Extended to catch Error and CIMError