Fix issues with the connection select subcommand.
"connection select" subcommand was failing to change the connection.

The changes included removing some bad code in _context_obj that
should have been removed some time ago and was causing the old
ContextObj to always be used and learning how the Click parent works.

It turns out that for a typical subcommand you are several layers down
in click context parent child relations.  We modified the subcommand
connection select to change the context.obj for the whole parent tree.

Other minor changes include:

1. _common.py - Change a raise ValueError to raise click.ClickException
for the abort of the pick operation.

2. Modify the __repr__ and __str__ in pywbem_server.  Since I had to use
them myself for this test, I found this to be a better representation.

3. Extend some documentation in the code.

4. _connection_repository.py, save did not properly delete the file when
it was empty.

5. execute_cmd did not test for existence of conn before trying to use
it.

6. Pywbem_server did not put the log parameter into the dictionary for
the json writes.

7. Corrects several issued found with new flake8. The new flake8 flags
lines with leading or trailing binary operator. Instances of this were
either corrected or # noqa hidden.

In addition, I completely rewrote the test_connection.py code to use
pytest and also to add more tests.