Change development support files to to match pywbem. These changes make keeping pywbemtools and pywbem development environments similar.
1. Change .gitconfig .flake8, .travis.yml to match pywbem

2. Change LICENSE to LICENSE.txt to match pywbem

3. Modify much of Makefile to match pywbem. This is an extensive set of
changes but tries to match the latest pywbem makefile as close as
possible. NOTE: This includes bypass for pywbem_os_setup.sh download
with cygwin because of issue. Will remove when we sort out issue with
cygwin.

4. Modify README.rst to correct error in license definition

5. Add README_PYPI.rst file using same pattern as pywbem

6. Modify appveyor.yml to match pattern set by pywbem

7. Modify dev-requirements to remove Jupyter requirement

8. Modify setup.cfg to match pywbem

9. Modify test_pegasus to correct flake8 error

10 Modify test_class_subcommand to add a test

11. Modify test_helpmsgs.py to fix issue with win32

12. Modify spelling error in test_instance_subcmd.py

13. Remove test_pywbemclicmds.py.  It was redundant.

14. Modify utils.py to fix issue with windows ci. For some reason the
tests faile with \r\r\n when Popen used to execute pywbemcli. This
is apparently an issue that is the latest Click.

15. Add methods from pywbem in tools to match pywbem and the changes
in the Makefile.  This includes package_version.py, python_bitsize.py,
python_version.py, and retry.bat.

16. Modify tox.ini to match pywbem

17. Commit MANIFEST.in

Fix tests because of windows

Click in windows sometimes formats help slightly different than in linux
systems.  We fixed this by creating a new test type that removes all
whitespace from the test and actual values and then tests the remainder
for equality.

Fix minimum-constraints to match pywbem

Create tmp pywbem_os_setup_cygwin.sh for cygwin to fix issue
define Makefile platform and env targets consistent with pywbem