Rewrite the methods that persist the repo of connections
This was a code cleanup to create a class for managing the repository
of connections in  place of prototype separate functions.  It added the class
ConnectionRepository and removed the separate methods

In addition to the changes to the connection_repository file it changed:

1. pywbemcli.py to use this class amd allow creating a new repository
from the command line.

2. the connection subcommands _cmd_connection.py to:

  a. allow the user to verify operations that change the repository
before they are executed.

  b. save the current server definition to the repository (connection
save).

  c. Rename the command that adds new connections to add.

3. Move the verify_option from _cmd_instances.py to _common_options.py.

4. Modify the name of the connection file from CONNECTION_FILE to
DEFAULT_CONNECTION_FILE under the assumption that in the future we will
allow user selected file names.

5. Modify the  tests tests_connection_repository.py to add tests.

6. Modify one method in _common (pick_from_list) so that
there is an internal _pick_from_list() and the pick_from_list external where theexternal returns the selected value, not just an index to that value.

7. Removed the test file test_connection_repository.py because is was
just a duplicate of the tests now in test_connection.py and did not
contribute anything to the test of pywbemcli. This will be
replaced in the future with a valid unit test on the
ConnectionRepository class.