Correct issue in log option usage.
pywbemcli was not passing log option to server object when name was used
as the definition of the server or when the connection select was used
to select another server from the repository.  This required adding log
option to the ContextObj so it is available for subsequent operations in
a repl session.

We also found a couple of places where we using the print output when it
should have been click.echo. changed.

Finally, I found that displaying individual objects such as paths was
difficult to read because there was no space between the objects. We
added a space and also a space after the last object.

Minor change to Makefile py.test to match pywbem makefile exactly.

Changed test files to match the change in display of paths, etc.