Fix a number of test issues resulting primarily from changes to the mocker exception messages and operation processing (more tests on input validity).  All of the changes here execpt for the one involving an import to config.py which was just my booboo are things found when we moved the testing of pywbemcli forward from pywbem 0.12.0.
Fix tests for CIMError message texts in tests.

The only part of these tests that is really important is
a) we got an exception.
b) it should match the defined error type although since this is
data just  from the mocker, it may not be the same from a real server.
The text for a number of the CIMError messages changed in pywbem 0.13.0
so we reduced the test to just confirming exception and exception status
code.

Removed poorly done global in config.py

This variable only was half defined so we removed it for now.  Without
this change the CI tests fail.

Add verbose input parameter to execute_pywbemcli() test function.

Change test that does not work.

One test in globals fails because it tries to build a class where no
qualifier decls exist in repo. This worked in earlier versions of pywbem
because the mocker was not checking for qualifier decls in CreateClass.
Changed the test to execute a qualifier list rather than class list.

Insert updates docs/helpmsg to match what is in help messages.

Modify url for os_setup files in Makefile. The make develop was failing
because the urls had changed in pywbem releases.

The urls for these shell scripts has changed since the release of pywbem
0.12.0.  First, the url for the read-the-docs url now includes a guid so
that does not seem too stable.

Change the makefile to use the url from github per the pywbem
documentation.

Fix results test for test_log_option.

We no longer put the pywbemcli at the beginning of the mock wbem
connection name.

Update the requirements.txt to reflect 0.13.0 as the min requirement

This simplifies the testing of mock returns.

Modified one test in _common.py that tests for output_format. The original
code was failing flake8.

Found one error in code in _common.py because a print statement did not
include the value to be printed.