Initial changes to sync pywbemtools setup to pywbem setup
Reduce number of differences in the setup  and install files between
pywbem and pywbemtools.  Goal is to make it easier to keep the projects
in sync.

NOTE: At this point the appveyor plaform bypasses the tests component.
That is another pr.

This set of changes should make the following file look as much as
possible like the same files in pywbem.

1. .flake8 - change to match pywbem .gitignore

2. .gitignore - change to match to pywbem except required differences

3. .travis.yml - match to pywbem

4. dev-requirements- Remove unneeded dev requirement for Jupyter. the
cli does not use Jupyter because the bash kernel for Jupyter was not
usable.

5. LICENSE.txt - Change name tofrom LICENSE to LICENSE.txt just to
reduce number of differences with pywbem

6. README.rst - Fix bad uri in README.rst so it points to correct file

7. README_PYPI.rst - Add README_PYPI.rst similar to pywbem.

8. setup.cfg - Add some lines (project_urls, description-file) to match
pywbem.  This uses READEME_PYPI.rst.

7. tests/live-unit/test_pegasus.py - Fix flake8 issue (extra space) in
code.

8. Makefile - Change the Makefile to match the pattern that now exists
for the pywbem makefile including targets for platfrom, env, inclusion
of version in target definitions, etc. This was extensive changes since
many of the targets and patterns for processing targets changed.

9. Remove the test file tests/unit/test_pywbemclicmds.py. It is obsolete
and completely replaced by the tests for each command.

10. Add the pywbem tools that are used now in pywbem as part of the makefile
and for .travis.yml and appveyor. These tools are copied directly with
no functional changes. The tools are retry.bat, package_version.py,
python_bitsize.py, python_version.py

11. Minimum-constraints.txt - update the constraint definitions.

12. Modify appveyor.yml to match the appveyor.yml defined in pywbem.

12. Commit MANIFEST.in to github.