Fix tests for differences between windows and other platforms
There are several differences in pywbemcli output between windows and
other platforms including:

1. The pywbemcli output may be formatted slightly differently on windows
than on other platforms, in particular different line formatting for the
help output text. Fixed by adding a test "linesnows" that compares the
actual and expected responses after eliminating all whitespace.  This is
now used for all of the tests of the help command insuring that the text
is the same but ignoring whitespace.

2. We cannot use the export command for windows and so had to modify
some tests to use set vs export.  Modified code to use set on windows

3. Modified tox.ini to reenable windows tests.