Add support for showing associations as a shrub
Add a new pywbem instance command(shrub), displaying a shrub where
a shrub is a view of the complete association relationship of a
particular cim instance showing not only the reference and
association information but the path of Role, AssocClass, ResultClass,
and ResultRole that formed the chain from the source instance
to the Result instances. It can display the complete set of
relationships or be limited with parameters that are filters to
particular sets of roles, association classes, and result roles.

The result by default is displayed as a tree in the following form

TST_Person.name="Mike"
 +-- member
 |   +-- TST_MemberOfFamilyCollection
 |       +-- family
 |           +-- TST_FamilyCollection(1 insts)
 |               +-- TST_FamilyCollection.name="Family2"
 +-- parent
     +-- TST_Lineage
         +-- child
             +-- TST_Person(2 insts)
                 +-- TST_Person.name="Sofi"
                 +-- TST_Person.name="Gabi"

The shrub may also display a table fo the results or the set
of objects that make up the result (MOF, etc.)

This also includes the capability to shorten the instance
paths displayed by changing selected key/value pairs to the
single character ~ if a) the keys are the same for all associated
instance paths or b) if they are certain known keys including
SystemCreationClassName, CreationClassName, or SystemName under
the assumption that these are all well known and do not really
contribute to separating the instance paths.  This is defaulted
to enabled but may be disabled with the option --no-fullpath.

New Class - Creates a new file and new class, AssociationShrub that
encapsulates all of the code to gather data and present the shrub as
either a table or an ascii tree.

Add complex association model for tests

Adds a ternary association model for shrub testing(complex_assoc).

Extends the documentation to include the new shrub command.