Add tests to test_globaloptions.py
Added a number of tests to test_globaloptions.py for most
global parameters both valid and invalid input options. Also removed the
original unittest tests and integrated them into pytest test.

In the process of this I found the following errors:

1. We were not validing the existence of the mock files until a
connection was made. Seems that the file mock file
existence should be validated at startup. Note that this does not
validate the data since that requires compiling, etc and we put that off
until we create the connection.

2. We were not aborting when a mock file was bad. That is a serious
enough error to warrent aborting.  Added click.Abort()

3. Missing parameter displayed for connection show (pull_max_count).
Also modified format slightly for connection show

4. Removed extraneous print statements from earlier commit.

5. Fixed other test in accord with new pywbemcli changes. The changes
that were incompatible are:

    a. Tests that use connection show output to determine state. This causes
       changes to test_connection_subcmd.py

6. Remove several worthless comments from code.