Fix params
Modified the input parameters for connection add.  Originally it used
two arguments (both required) for server and name. With the number of
options on this subcommand it was easy to lose the two arguments.
Besides we should do this parallel to the globals which used options for
these input parameters so it was logical to modify this to use --server
and --name also.

Also fixed tests.