Clean up the includequalifiers option on instance enumerate, associators, and references subcommands.
While these options are defined for these operations, pywbemcli may use
either the pull operations or the traditional (EnumerateInstances, etc)
operations. Thus the behavior if "insance enumerate", etc is
dependent on the pywbem API actually called through the Iter... pywbem
API.  If the pull operations are used, they are not used since the
IncludeQualifiers is not part of this operation; if the traditional
operations are used, the server may or may not return qualifiers.

We also modified LocalOnly documentation for the instance enumerate
because of some of the same issues.

We changed the help for these operations to reflect the possibilities,
and issues in their use and added tests to confirm that executing these
subcommands no longer fails.

We also included changed the word global to general for the general option in
_cmd_class documentation. pywbemcli help shows the "GENERAL" options be
when the project started, we called them global which could confuse the
reader.

We added a number of tests of includequalifiers and localhost
to assure that they are executing.

Corrected bad wording in the --names-only option.