Fix class find subcommand
The help was incorrect for this subcommand because it used the standard
namespace option but the behavior is different. The standard option is
about selecting another namespace than the default.  This is about
limiting the search to a namespace.

At the same time it seemed logical to extend it to allow using multiple
namespaces rather than a single namespace with the option so we made that
change also.