Remove now worthless test
test_helpmsgs.py was written early in the development of pywbemtools
just to go through all the help messages.  It is out of date and
superceeded by the fact that we should be testing all help messages in
each command-group test.

Since it wastes time in the tests for no reason I deleted it.