Modify connection table
Modify the connection table so that it only shows rows that are not
empty. Since many of the possible rows of connections may be empty, this
makes the resulting display much more compact and readable.

Also remove the password display from this table.  The password is still
visible from the connection show command.