Add FilterQuery and FilterQueryLanguage options
The options to define the FilterQuery and FilterQueryLanguage were
missing from the subcommand definitions for associators, references,
enumerateinstances.

The code exists correct in the lower layers but the options to actually
build them were missing.

Remember that we use the Iter... operations for all of these so these
parameters should have been there so that when pull operations are
executed by pywbemcli_operations, those parameters can be made
available.

NOTE: We ignore the fact that these parameters may have been set by the
user and pywbemcli uses the original operations (i.e.
EnumerateInstances) where these parameters are not passed to the server;
we do not stop the operation or warn the user.  Since it is not
pywbemcli that always controls whether pull or original operations are
used, we cannot completely determine whether pywbem will use the pull or
original operations except in the case where the --use-pull-ops flag is
set to force the original operations. Thus we ignore the issue and leave
it up to the server.

Also, the pywbemcli behavior is different than the server behavior with
respect to the FilterQueryLanguage option.  If FilterQuery exists and
FilterQueryLanguage does not exist, most servers reject the operation
request. Pywbemcli substitutes "DMTF:FQL" into QueryFilterLanguage
since that is the logical choice.